// src/domain/enum/user.enum.ts
export enum UserTypeEnum {
  INDIVIDUAL = "individual",
  BUSINESS = "business",
}

export enum RoleTypeEnum {
  ADMIN = "admin",
  USER = "user",
}

export enum GenderEnum {
  MALE = "male",
  FEMALE = "female",
}

export enum AuthTypeEnum {
  GOOGLE = "google",
  FACEBOOK = "facebook",
  APPLE = "apple",
  LOCAL = "local",
}

export enum FollowTypeEnum {
  FOLLOWERS = "followers",
  FOLLOWINGS = "followings",
  REQUESTED = "requested",
  NOT_FOLLOWING = "not-following",
}

export enum FollowStatusTypeEnum {
  FOLLOW = "follow",
  FOLLOWING = "following",
  REQUESTED = "requested",
}

export enum MediaTypeEnum {
  PHOTOS = "photos",
  VIDEO = "video",
  TEXT = "text",
  AUDIO = "audio",
  FONT = "font",
  EBOOK = "e-book",
  PDF = "pdf",
}